Revert "Attempt to make properties with physical meaning"
This reverts commit aab497b1937de439b65c552ee9f714887471780e.

+++ DeviceAdapters/MicroFPGA/MicroFPGA.h
@@ -104,10 +104,10 @@ private:
 	bool initialized_;
 	bool mode_;
 	bool start_;
-	double pulse_;
-	double period_;
-	double exposure_;
-	double delay_;
+	long pulse_;
+	long period_;
+	long exposure_;
+	long delay_;
 	bool busy_;
 };
